% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bsplines}
\alias{bsplines}
\title{Bsplines- Computes the functional spatial median of the functional data}
\usage{
bsplines(fdaobj, norder, lambda, Lf)
}
\arguments{
\item{fdaobj}{An object of the FDA Class.
#'}

\item{fdaobj}{An object of the FDA Class.
#'}

\item{fdaobj}{An object of the FDA Class.
#'}

\item{fdaobj}{An object of the FDA Class.}
}
\value{
A component list with:
\itemize{
\item{"argvals "}{}
\item{"fdSmooth"}{}
\item{"df"}{}
\item{"SSE"}{}
\item{"PenMat"}{}
\item{"GCV"}{}
}
}
\description{
Bsplines- Computes the functional spatial median of the functional data
}
\examples{
\dontrun{
norder<-6
lambda <- 0.01
Lf<-4
fdaobj<-fdaobjMale
BSplinesGrowth <-bsplines( fdaobj, norder, lambda, Lf )
matplot(BSplinesGrowth$fdSmooth$coefs, type="l",main="Smooth Data using B-Splines",xlab="Age (years)", ylab="Heigth (cm)")
}
}
\references{
Ramsay, James O., and Silverman, Bernard W. (2006), Functional Data Analysis, 2nd ed., Springer, New York.
             Ramsay, James O., and Silverman, Bernard W. (2002), Applied Functional Data Analysis, Springer,New York.
}

